feat: add NodeStatus component for displaying and managing node status
